Pass Subject instance to grid sampler and aggregator (#129)
* Add property to get shape of first image in sample

* Fix typing hint

* Pass subject sample to grid sampler and aggregator

Resolves #100:

The output of a `GridAggregator` was hard-coded to have a single
channel. However, users should be able to aggregate inference results
with multiple channels. An out_channels argument has been added to
customize the aggregator behavior.

Resolves #127:

The grid sampler and aggregator worked with tensor directly, which was
limiting. They now take instances of `Subject` and guess the shape from
the first image in the sample.

* Update inference classes docs